fix(review): address Copilot feedback for feed endpoint and test structure
- Shifted DB query construction inside format.xml block to avoid running database requests on unsupported formats.
- Used `head :not_acceptable` explicitly for all other formats, avoiding `ActionController::UnknownFormat` exceptions and keeping 406 responses performant.
- Moved testing code from `home_controller_test.rb` to dedicated `test/controllers/home/feed_test.rb` following the project's TEST_ARCHITECTURE.md per-action file convention.
