+++ app/controllers/home_controller.rb
@@ -122,14 +122,16 @@ class HomeController < ApplicationController
   def feed
     ahoy.track "Rendered feed"
 
-    @events = Event.includes([:country, [uploads_attachments: :blob]])
-      .venontaj
-      .ne_nuligitaj
-      .where(cancelled: false)
-      .order(:date_start)
-
     respond_to do |format|
-      format.xml { render layout: false }
+      format.xml do
+        @events = Event.includes([:country, [uploads_attachments: :blob]])
+          .venontaj
+          .ne_nuligitaj
+          .where(cancelled: false)
+          .order(:date_start)
+        render layout: false
+      end
+      format.all { head :not_acceptable }
     end
   end
